# chore: Improve benchmark output