**What this PR does / why we need it**:
Simply adds a suffix to the filenames so benchmarks will look like this and make it easier to know what to compare with each other.

```
pkg: k8s.io/kube-state-metrics/v2/pkg/app
                                   │ /tmp/tmp.eoDSvm8Gwf-main │         /tmp/tmp.tKQqwKPsmD-mybranch         │
                                   │       sec/op        │   sec/op     vs base                │
KubeStateMetrics/GenerateMetrics-4           1.001 ± ∞ ¹   1.001 ± ∞ ¹       ~ (p=1.000 n=1) ²
KubeStateMetrics/MakeRequests-4              7.805 ± ∞ ¹   7.627 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                      2.795         2.763        -1.15%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```


**How does this change affect the cardinality of KSM**: *(increases, decreases or does not change cardinality)*
None
**Which issue(s) this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when PR gets merged)*:
Fixes #